Nacos

+++ taroco-rbac/src/main/java/cn/taroco/rbac/admin/service/impl/SysDeptServiceImpl.java
@@ -1,16 +1,18 @@
 package cn.taroco.rbac.admin.service.impl;
 
-import cn.taroco.rbac.admin.mapper.SysDeptMapper;
-import cn.taroco.rbac.admin.model.entity.SysDeptRelation;
-import com.baomidou.mybatisplus.mapper.EntityWrapper;
-import com.baomidou.mybatisplus.service.impl.ServiceImpl;
+import cn.taroco.common.constants.CommonConstant;
+import cn.taroco.common.exception.BusinessException;
 import cn.taroco.rbac.admin.common.util.TreeUtil;
+import cn.taroco.rbac.admin.mapper.SysDeptMapper;
 import cn.taroco.rbac.admin.mapper.SysDeptRelationMapper;
 import cn.taroco.rbac.admin.model.dto.DeptTree;
 import cn.taroco.rbac.admin.model.entity.SysDept;
+import cn.taroco.rbac.admin.model.entity.SysDeptRelation;
+import cn.taroco.rbac.admin.model.entity.SysRoleDept;
 import cn.taroco.rbac.admin.service.SysDeptService;
-import cn.taroco.common.constants.CommonConstant;
-import org.springframework.beans.BeanUtils;
+import cn.taroco.rbac.admin.service.SysRoleDeptService;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,8 +32,12 @@ import java.util.List;
 public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> implements SysDeptService {
     @Autowired
     private SysDeptMapper sysDeptMapper;
+
     @Autowired
     private SysDeptRelationMapper sysDeptRelationMapper;
+
+    @Autowired
+    private SysRoleDeptService sysRoleDeptService;
 
     /**
      * 添加信息部门
@@ -41,10 +47,8 @@ public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> impl
      */
     @Override
     public Boolean insertDept(SysDept dept) {
-        SysDept sysDept = new SysDept();
-        BeanUtils.copyProperties(dept, sysDept);
-        this.insert(sysDept);
-        this.insertDeptRelation(sysDept);
+        this.save(dept);
+        this.insertDeptRelation(dept);
         return Boolean.TRUE;
     }
 
@@ -56,7 +60,7 @@ public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> impl
         //增加部门关系表
         SysDeptRelation deptRelation = new SysDeptRelation();
         deptRelation.setDescendant(sysDept.getParentId());
-        List<SysDeptRelation> deptRelationList = sysDeptRelationMapper.selectList(new EntityWrapper<>(deptRelation));
+        List<SysDeptRelation> deptRelationList = sysDeptRelationMapper.selectList(new QueryWrapper<>(deptRelation));
         for (SysDeptRelation sysDeptRelation : deptRelationList) {
             sysDeptRelation.setDescendant(sysDept.getDeptId());
             sysDeptRelationMapper.insert(sysDeptRelation);
@@ -76,11 +80,18 @@ public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> impl
      */
     @Override
     public Boolean deleteDeptById(Integer id) {
+        // 删除前判断部门下面是否包含角色
+        final SysRoleDept sysRoleDept = new SysRoleDept();
+        sysRoleDept.setDeptId(id);
+        final int count = sysRoleDeptService.count(new QueryWrapper<>(sysRoleDept));
+        if (count > 0) {
+            throw new BusinessException("部门下尚存在角色, 不能删除");
+        }
         SysDept sysDept = new SysDept();
         sysDept.setDeptId(id);
         sysDept.setUpdateTime(new Date());
         sysDept.setDelFlag(CommonConstant.STATUS_DEL);
-        this.deleteById(sysDept);
+        this.updateById(sysDept);
         sysDeptMapper.deleteDeptRealtion(id);
         return Boolean.TRUE;
     }
@@ -109,8 +120,8 @@ public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> impl
      * @return 树
      */
     @Override
-    public List<DeptTree> selectListTree(EntityWrapper<SysDept> sysDeptEntityWrapper) {
-        return getDeptTree(this.selectList(sysDeptEntityWrapper), 0);
+    public List<DeptTree> selectListTree(QueryWrapper<SysDept> sysDeptEntityWrapper) {
+        return getDeptTree(this.list(sysDeptEntityWrapper), 0);
     }
 
     /**
@@ -131,6 +142,7 @@ public class SysDeptServiceImpl extends ServiceImpl<SysDeptMapper, SysDept> impl
             node.setId(dept.getDeptId());
             node.setParentId(dept.getParentId());
             node.setName(dept.getName());
+            node.setOrderNum(dept.getOrderNum());
             trees.add(node);
         }
         return TreeUtil.bulid(trees, root);
